+++ test_util/src/test_server.rs
@@ -1,5 +1,5 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
 fn main() {
-  test_util::run_all_servers();
+  test_util::servers::run_all_servers();
 }
